requirements.txt
asgiref==3.5.2
attrs==22.1.0
colorama==0.4.5
coverage==6.4.4
Django==4.1.1
djangorestframework==3.14.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
